Cost: base_score documentation

+++ src/Cost.h
@@ -14,11 +14,12 @@ class Cost {
 
         CUDAFLAGS static int base_score(const char &a, const char &b); //'alpha' score
         CUDAFLAGS static float match_score(const char &a, const char &b); // 'beta' score
-        CUDAFLAGS static int compensation_score(const char &a1, const char &a2, const char &b1, const char &b2); // 'tau' score
+        CUDAFLAGS static float compensation_score(const char &a1, const char &a2, const char &b1, const char &b2); // 'tau' score
 };
 
 CUDAFLAGS inline int Cost::base_score(const char &a, const char &b)
 {
+    /* In the BP scores, this function only checks if it is a base pair */
     if ((a == 'C' && b == 'G') ||
         (a == 'G' && b == 'C') ||
         (a == 'A' && b == 'U') ||
@@ -38,7 +39,7 @@ CUDAFLAGS inline float Cost::match_score(const char &a, const char &b)
     return 0;
 }
 
-CUDAFLAGS inline int Cost::compensation_score(const char &a1, const char &a2, const char &b1, const char &b2)
+CUDAFLAGS inline float Cost::compensation_score(const char &a1, const char &a2, const char &b1, const char &b2)
 {
     // Give a score if they have the same structure
     if (base_score(a1, a2) == base_score(b1, b2))
